+++ cmd/conven/main.go
@@ -10,8 +10,8 @@ import (
 	"github.com/leo1394/homebrew-conven/internal/cli"
 )
 
-var version = "1.0.3"
-var versionDate = "2026-09-07"
+var version = "1.0.4"
+var versionDate = "2026-09-11"
 
 func main() {
 	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)